OLED improvement: Implemented diagram selection at the tree.

+++ br.ufes.inf.nemo.oled/src/br/ufes/inf/nemo/oled/ui/TreePopupMenu.java
@@ -9,6 +9,8 @@ import javax.swing.JPopupMenu;
 import javax.swing.tree.DefaultMutableTreeNode;
 import javax.swing.tree.TreePath;
 
+import br.ufes.inf.nemo.oled.umldraw.structure.StructureDiagram;
+
 public class TreePopupMenu extends JPopupMenu {
  
 	private static final long serialVersionUID = 1L;
@@ -17,13 +19,16 @@ public class TreePopupMenu extends JPopupMenu {
 	public JMenuItem autoCompleteItem = new JMenuItem("Complete Selection...");
 	public JMenuItem refreshItem = new JMenuItem("Refresh");
 	
-    public TreePopupMenu(final AppFrame frame, final ProjectTree tree)
+	public Object element;
+	
+    public TreePopupMenu(final AppFrame frame, final ProjectTree tree, Object element)
     {       
+    	this.element = element;    	
     	DefaultMutableTreeNode node = ((DefaultMutableTreeNode)tree.getSelectionPath().getLastPathComponent());
+    	
+    	// Auto Complete Selection
     	if (tree.getModelRootNode().equals(node)){
-	    	add(autoCompleteItem);
-	    	addSeparator();
-	    	
+	    	add(autoCompleteItem);	    	
 	    	autoCompleteItem.addActionListener(new ActionListener() {				
 				@Override
 				public void actionPerformed(ActionEvent e) {
@@ -33,33 +38,43 @@ public class TreePopupMenu extends JPopupMenu {
 			});
 	    }
     	
-    	add(refreshItem);
-    	addSeparator();
+    	// Refresh Tree
+    	if (tree.getRootNode().equals(node)){
+    		add(refreshItem);
+    		refreshItem.addActionListener(new ActionListener() {				
+    			@Override
+    			public void actionPerformed(ActionEvent e) {
+    				//FIXME every modification creates a new tree
+    				ProjectBrowser.rebuildTree(frame.getDiagramManager().getCurrentProject());
+    			}
+    		});
+    	}    	
     	
-    	refreshItem.addActionListener(new ActionListener() {				
-			@Override
-			public void actionPerformed(ActionEvent e) {
-				//FIXME every modification creates a new tree
-				ProjectBrowser.rebuildTree(frame.getDiagramManager().getCurrentProject());
-			}
-		});
+    	// Delete 
+    	if (!tree.getModelRootNode().equals(node) && !tree.getDiagramRootNode().equals(node) && !tree.getRootNode().equals(node)){    		
+    		add(deleteItem);
+    		deleteItem.setIcon(new ImageIcon(TreePopupMenu.class.getResource("/resources/br/ufes/inf/nemo/oled/ui/delete.png")));
+    		deleteItem.addActionListener(new ActionListener() {				
+    			@Override
+    			public void actionPerformed(ActionEvent e) {			
+    							
+    				if (TreePopupMenu.this.element instanceof OntoUMLElement)
+    				{
+    					OntoUMLElement ontoElem = (OntoUMLElement) ((DefaultMutableTreeNode)tree.getSelectionPath().getLastPathComponent()).getUserObject();
+    					frame.getDiagramManager().deleteElementOfProject((RefOntoUML.Element)ontoElem.getElement());													
+    					// FIXME every modification creates a new tree
+    					ProjectBrowser.rebuildTree(frame.getDiagramManager().getCurrentProject());					
+    					tree.setSelectionPath(new TreePath(tree.getModelRootNode().getPath()));
+    				}else if (TreePopupMenu.this.element instanceof StructureDiagram)
+    				{
+    					frame.getDiagramManager().deleteDiagram((StructureDiagram)TreePopupMenu.this.element);
+    					// FIXME every modification creates a new tree
+    					ProjectBrowser.rebuildTree(frame.getDiagramManager().getCurrentProject());
+    				}
+    			}
+    		});
+    	}   	    	
     	
-    	add(deleteItem);
-    	deleteItem.setIcon(new ImageIcon(TreePopupMenu.class.getResource("/resources/br/ufes/inf/nemo/oled/ui/delete.png")));
-    	    	
-    	deleteItem.addActionListener(new ActionListener() {				
-			@Override
-			public void actionPerformed(ActionEvent e) {			
-												
-				OntoUMLElement ontoElem = (OntoUMLElement) ((DefaultMutableTreeNode)tree.getSelectionPath().getLastPathComponent()).getUserObject();
-				frame.getDiagramManager().deleteElementOfProject((RefOntoUML.Element)ontoElem.getElement());
-												
-				// FIXME every modification creates a new tree
-				ProjectBrowser.rebuildTree(frame.getDiagramManager().getCurrentProject());
-				
-				tree.setSelectionPath(new TreePath(tree.getModelRootNode().getPath()));
-			}
-		});
     }
